requirements.txt
icecream==2.1.1
pandas==1.3.0
scikit-allel==1.3.5
